cleanup and fix bug in loading
