+++ app/game.tsx
@@ -7,9 +7,24 @@ import { getOrUpdateStatus } from "@/functions/getOrUpdateStatus";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import { useEffect, useState } from "react";
 import Loading from "@/components/Loading";
+import { getQuestion } from "@/functions/getQuestion";
 
 export default function Game() {
   const [status, setStatus] = useState<string>("");
+  const [question, setQuestion] = useState<string>("");
+
+  const fetchQuestion = async () => {
+    const gameRoom = await AsyncStorage.getItem("gameRoom");
+    if (gameRoom) {
+      const unsubscribe = await getQuestion(gameRoom, setQuestion, undefined);
+
+      return () => {
+        if (unsubscribe) {
+          unsubscribe();
+        }
+      };
+    }
+  };
 
   useEffect(() => {
     const ListenAndGetStatus = async () => {
@@ -27,31 +42,23 @@ export default function Game() {
       }
     };
     ListenAndGetStatus();
+    fetchQuestion();
   }, []);
 
   return (
     <ParallaxScrollView>
-      {status === "waiting" ? (
-        <Loading />
-      ) : (
+      {!!status && (
         <>
-          <PlayerIcon size={80} />
-          <AddAnswer />
+          {status === "waiting" ? (
+            <Loading />
+          ) : (
+            <>
+              <PlayerIcon size={80} />
+              <AddAnswer question={question} />
+            </>
+          )}
         </>
       )}
     </ParallaxScrollView>
   );
 }
-
-const styles = StyleSheet.create({
-  headerImage: {
-    color: "#808080",
-    bottom: -90,
-    left: -35,
-    position: "absolute",
-  },
-  titleContainer: {
-    flexDirection: "row",
-    gap: 8,
-  },
-});
