+++ components/ButtonComponent.tsx
@@ -1,5 +1,5 @@
 import { Colors, Sizes } from "@/constants/Theme";
-import { StyleSheet, TouchableOpacity, Text } from "react-native";
+import { StyleSheet, Pressable } from "react-native";
 import { ThemedText } from "./ThemedText";
 import { Href, useRouter } from "expo-router";
 
@@ -33,14 +33,14 @@ export function ButtonComponent({
   };
 
   return (
-    <TouchableOpacity
+    <Pressable
       style={[styles.button, { backgroundColor: buttonColor }]}
       onPress={handlePress}
     >
       <ThemedText style={styles.buttonText} type="defaultSemiBold">
         {text}
       </ThemedText>
-    </TouchableOpacity>
+    </Pressable>
   );
 }
